Made frog finally walk/ added run animation that needs transition on the blend trees/ and made the net move with it
This is all within the animations folder where I made a frog folder to hold anything related to the frog animations/ tpose rig/ and the animation controller.  Took me a while due to for whatever reason the animations were not being read although when I made it work I followed the same steps as when it did not work.   I really don't understand what I was missing, but hey its working now hopefully.