Update on frog Animations and Animator Controller
Push updating the animator controller showcasing all the animations up to date in the frog and delete the unnecessary animations that were in the hierarchy.  The blend trees are still rough and can be edited but it at least loops through all of Aracelys animations.  Also turned off fixed exit time and fixed duration.