Fix textures of the 3 bugs(Beetle, Firefly, Moth)
Had to re unwrap the models since they were not matching up with the original textures.  This means the current bug models in unity within the bug spawner need to be changed with the the newly textured models I added as prefabs in unity.  The newly updated prefabs are within the folders Assets>Prefabs>Bugs<FixedModelTextures.  I tried inputting it into the spawner system but I could not get it to work, but the new model textures are there.