Cinemachine transition added with a MainStory Script
Experimented trying to add the intro cutscene.  It works for the most part, only thing is I would maybe try to put the loading screen after the cutscene since it takes a while to load and looks like it freezes(But its really just loading the next scene).  I was trying to figure out how to put the loading screen after the cutscene but I haven't been able to figure it out.  Either way it kind of works.