Bug Shoot Particle Effect test in new scene
Not sure if I'll end up making a particle effect or not for the bugs disappearing since I'm not sure yet how to code it into the other bugs, but I was able to make a somewhat workable effect.  I'll see if I can get it to work or not before thursday as I continue working on this as well as my other assigned tasks.  Also the effect is made from already existing assets we have from our game, I just put them out to the effect to see if it works or not.